Change CacheBuster.jsx to CacheBuster.tsx & add types
